Fixed front end and backend constants files to work
