+++ constants.py
@@ -1,0 +1,2 @@
+STATUS_SUFFIX = "CUDSTATUS"
+SEVERITY_SUFFIX = "CUDSEVR"
